Match the amplitude
SSB calculates a value that is proportional to the intensities.

By taking the square root of the calculated amplitude, we can derive an amplitude
that matches the reference very well.

This proves it works for small amplitude variations. To be validated
for larger variations as well -- could also be division by square root of average,
average of square root, square root of maximum, ...!