Branch between efficient single thread CPU and  multi-thread CPU version
The single thread CPU variant benefits from a transposed copy and is very efficient,
while the multithread CPU variant is significantly less efficient (25 % of single thread)
but achieves short start-to-finish times, probably since
it eliminates the transpose step.

An experimental Numba version that used threading on transposed
input data achieved 30 % longer start-to-finish time at about twice the
efficiency of the current multithread version. In theory it could be regarded as a
superior compromise, but it is unclear under which real-world circumstances
it would bring an advantage.